examples/generative/image_generation/repa/requirements.txt
# REPA dependencies - frozen versions for reproducibility
torch>=2.1.0
torchvision>=0.16.0
diffusers==0.27.2
transformers==4.40.0
timm==0.9.16
accelerate==0.29.3
huggingface_hub==0.22.2
safetensors==0.4.3
datasets==2.18.0
tqdm==4.66.1
numpy==1.26.4
Pillow==10.2.0
einops==0.7.0
requests==2.31.0
wandb==0.16.3

# CLIP (for optional CLIP encoder support)
ftfy==6.1.3
regex==2023.12.25
git+https://github.com/openai/CLIP.git
